added a code

+++ exercise_5.py
@@ -3,13 +3,13 @@
 # The program should then calculate the student's grade as a percentage, and print out a message indicating whether the student passed or failed the exam.
 
 # prompting the teacher to enter the number of questions on the exam
-number_of_questions = int(input("Enter the number of questions: "))
+number_of_questions = float(input("Enter the number of questions: "))
 
 # prompting the teacher to enter the number of questions the student answered correctly
-number_of_questions_answered_correctly = int(input("Enter the number of questions student answered correctly: "))
+number_of_questions_answered_correctly = float(input("Enter the number of questions student answered correctly: "))
 
 # prompting the teacher to enter the number of points each question is worth
-number_of_points_each_worth = int(input("Enter the points each question is worth: "))
+number_of_points_each_worth =float(input("Enter the points each question is worth: "))
 
 # calculating the total number of points the student scored
 total_number_of_points = (number_of_questions_answered_correctly * number_of_points_each_worth)
